chore : 엔드포인트 수정(컨트롤러 분리) #68
