+++ src/main/java/com/example/momogum/web/controller/UserProfileController/FollowController.java
@@ -9,16 +9,51 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/follow")
+@RequestMapping("/follows")
 @RequiredArgsConstructor
 @Tag(name = "Follow API",description = "팔로잉 팔로우 API")
 public class FollowController {
+
   private final FollowService followService;
+
+  /**
+   * 팔로잉, 팔로워 수 카운트
+   */
+
+  @Operation(summary = "팔로워/팔로잉 수 조회 API", description = "특정 유저의 팔로워 및 팔로잉 수를 조회합니다.")
+  @GetMapping("/{userId}/followCount")
+  public ApiResponse<FollowDTO.FollowStatsDTO> getFollowStats(@PathVariable Long userId) {
+    return ApiResponse.onSuccess(followService.getFollowStats(userId));
+  }
+
+  /**
+   * 팔로잉 목록 조회
+   */
+
+  @Operation(summary = "팔로잉 목록 조회 API", description = "유저의 팔로잉(내가 팔로우하는 사람들) 조회")
+  @GetMapping("/{userId}/search/following")
+  public ApiResponse<List<FollowDTO.FollowingResponseDTO>> getAllFollowing(@PathVariable Long userId) {
+
+    return ApiResponse.onSuccess(followService.getFollowings(userId));
+  }
+
+
+  /**
+   * 팔로워 목록 조회
+   */
+
+  @Operation(summary = "팔로워 목록 조회 API", description = "유저의 팔로워(나를 팔로우하는 사람들) 조회")
+  @GetMapping("/{userId}/search/followers")
+  public ApiResponse<List<FollowDTO.FollowerResponseDTO>> getAllFollowers(@PathVariable Long userId) {
+
+    return ApiResponse.onSuccess(followService.getFollowers(userId));
+  }
 
   /**
    * 닉네임 또는 이름으로 팔로잉하는 유저 검색
